Use distance from Ewald sphere to improve orientation finding/fitting
Implement new get_score function that uses shape vector, approximated
as a gaussian, to adjust the pixel intensity. Initial tests show that this
significantly improves the stability of the orientation refinement, as well
as the initial orientation search. It also helps visualization, because it
is easier to identify which reflections are significant (i.e. close
to the Ewald sphere). The score function itself is slightly slower as a result
(about 20-30%).